build: define arguments as `--enable-[term]` instead of `--disable-[term]`
The arguments were called `disable-{tests,bench,hardening}` but the
variable they set was `use_[term]`, which could lead to an inversion.

It appears that it is expected that developers declare enablement flags
and Autotools will generate their equivalent disablement flags but there's
no text to indicate that the same will happen vice-versa.