build: autodetect i?86 and arm as 32-bit
This is a problem introduced by skipping architecture specific flags
when building with optimizations disabled.

Neither i?86 nor arm contain the number 32 in it, making number based
matching unusable. But arm64 is 64-bit (though usually the recommended
term is aarch64) and so we must match targets with the number 64 first
before we assume anything with arm is 32-bit.